Refactor credit handling to use Double for importo and credito_residuo

+++ Progetto-Coffee-cApp/src/main/java/it/unipa/wsda/progettocoffeecapp/controller/cliente/RicaricaController.java
@@ -20,7 +20,7 @@ public class RicaricaController {
 
     @PostMapping("/ricarica")
     public String ricaricaCredito(@RequestParam Integer id_utente,
-                                  @RequestParam Integer importo,
+                                  @RequestParam Double importo,
                                   Model model) {
         
         utenteService.ricaricaCredito(id_utente, importo);
